put _sampled_eventstream_timer in the right place
