+++ src/sentry/eventstream/kafka/backend.py
@@ -1,20 +1,8 @@
 import logging
-import random
 import signal
 import uuid
-from contextlib import contextmanager
 from enum import Enum
-from typing import (
-    Any,
-    Generator,
-    Literal,
-    Mapping,
-    MutableMapping,
-    Optional,
-    Sequence,
-    Tuple,
-    Union,
-)
+from typing import Any, Literal, Mapping, MutableMapping, Optional, Sequence, Tuple, Union
 
 from arroyo import Topic, configure_metrics
 from arroyo.backends.kafka.configuration import build_kafka_consumer_configuration
@@ -39,23 +27,11 @@ from sentry.utils.kafka_config import (
 
 logger = logging.getLogger(__name__)
 
-_DURATION_METRIC = "eventstream.duration"
-
 
 class PostProcessForwarderType(str, Enum):
     ERRORS = "errors"
     TRANSACTIONS = "transactions"
     ISSUE_PLATFORM = "issue_platform"
-
-
-@contextmanager
-def _sampled_eventstream_timer(instance: str) -> Generator[None, None, None]:
-    record_metric = random.random() < 0.1
-    if record_metric is True:
-        with metrics.timer(_DURATION_METRIC, instance=instance):
-            yield
-    else:
-        yield
 
 
 class KafkaEventStream(SnubaProtocolEventStream):
